core/requirements.txt
python-binance
pandas
ccxt
fastapi
uvicorn
Pillow
scipy
matplotlib
flask
dash
plotly
scikit-learn
tradingpattern
colorama